feat: move pages folder to src folder
